refactor: API 호출 함수 파일 분리
코드 복잡성을 줄이기 위해 API 호출 함수 파일을 분리했습니다.
